backend/requirements.txt
# Dependências do Backend FastAPI
# ================================

# Framework Web
fastapi
uvicorn[standard]

# Base de Dados
sqlalchemy
aiosqlite

# Autenticação e Segurança
passlib[bcrypt]
bcrypt
python-jose[cryptography]
python-multipart

# Email
fastapi-mail
email-validator

# OAuth (Google, Facebook)
authlib
httpx
itsdangerous

# Configuração
python-dotenv
pydantic-settings

# Migrations (opcional, mas útil)
alembic